blscfg: add a space char when appending fields for variable expansion
The GRUB variables are expanded and replaced by their values before adding
menu entries, but they didn't include space characters after the values so
the result was not correct.

For the common case this wasn't a problem but it is if there are variables
that are part of the values of other variables.

Resolves: rhbz#1669252

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>